+++ build.gradle
@@ -18,7 +18,8 @@ allprojects {
         gdxVersion = '1.10.0'
         visuiVersion = '1.5.0'
         kryoVersion = '5.2.0'
-        junitVersion = '4.12'
+        junitVersion = '4.13.2'
+        mockitoVersion = '1.10.19'
         commonsIoVersion = '2.5'
         commonsLangVersion = '3.12.0'
         gltfVersion = 'master-SNAPSHOT'
@@ -43,6 +44,8 @@ project(":commons") {
         compile "com.github.mgsx-dev.gdx-gltf:gltf:$gltfVersion"
 
         testCompile "junit:junit:$junitVersion"
+
+        testCompile "org.mockito:mockito-all:$mockitoVersion"
     }
 }
 
@@ -73,7 +76,9 @@ project(":editor") {
         compile "com.kotcrab.vis:vis-ui:$visuiVersion"
         compile "com.esotericsoftware:kryo:$kryoVersion"
 
+        // tests
         testCompile "junit:junit:$junitVersion"
+        testCompile "org.mockito:mockito-all:$mockitoVersion"
     }
 }
